Update for 1.1.1
Added discord webhook URL and logic for Stats with new function - .ptd - Admin Only - posts the top 10 players and top 5 clans. Mostly to be used right before server shutdown and save file wipe. We have the .rs command for resetting stats, but this will post to discord for leaderboard purposes and uses a single webhook to do once. The same webhook can be used still, but this command should only be ran once due to the size of the post.